+++ admin/partials/vs-injector-admin-display.php
@@ -16,7 +16,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 }
 
 if ( ! current_user_can( 'manage_options' ) ) {
-	wp_die( esc_html__( 'Sorry, you are not allowed to manage options for this site.', 'vinoshipper' ) );
+	wp_die( esc_html__( 'Sorry, you are not allowed to manage options for this site.', 'vinoshipper-injector' ) );
 }
 
 ?>
@@ -30,7 +30,7 @@ if ( ! current_user_can( 'manage_options' ) ) {
 	do_settings_sections( 'vs_injector_settings_section_general' );
 	do_settings_sections( 'vs_injector_settings_section_theme' );
 	do_settings_sections( 'vs_injector_settings_section_cart' );
-	submit_button( ( esc_html__( 'Save Settings', 'vinoshipper' ) ) );
+	submit_button( ( esc_html__( 'Save Settings', 'vinoshipper-injector' ) ) );
 	?>
 	</form>
 </div>